requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Flask>=2.0.0
Pillow>=8.0.0
